+++ scripts/corgos_telegram_bot/settings.py
